windows cookbook from opscode-cookbooks @ 767b997

+++ cookbooks/windows/libraries/registry_helper.rb
@@ -1,7 +1,7 @@
 #
 # Author:: Doug MacEachern (<dougm@vmware.com>)
 # Author:: Seth Chisamore (<schisamo@opscode.com>)
-# Author:: Paul Morotn (<pmorton@biaprotect.com>)
+# Author:: Paul Morton (<pmorton@biaprotect.com>)
 # Cookbook Name:: windows
 # Provider:: registry
 #
@@ -57,9 +57,9 @@ module Windows
       hkey = get_hive_name(path)
 
       hive = {
-        "HKEY_LOCAL_MACHINE" => Win32::Registry::HKEY_LOCAL_MACHINE,
-        "HKEY_USERS" => Win32::Registry::HKEY_USERS,
-        "HKEY_CURRENT_USER" => Win32::Registry::HKEY_CURRENT_USER
+        "HKEY_LOCAL_MACHINE" => ::Win32::Registry::HKEY_LOCAL_MACHINE,
+        "HKEY_USERS" => ::Win32::Registry::HKEY_USERS,
+        "HKEY_CURRENT_USER" => ::Win32::Registry::HKEY_CURRENT_USER
         }[hkey]
 
       unless hive
@@ -73,7 +73,7 @@ module Windows
 
     def unload_hive(path)
       hive = get_hive(path)
-      if hive == Win32::Registry::HKEY_USERS
+      if hive == ::Win32::Registry::HKEY_USERS
         reg_path = path.split("\\")
         priv = Chef::WindowsPrivileged.new
         begin
@@ -83,17 +83,18 @@ module Windows
       end
     end
 
-    def set_value(mode,path,values)
+    def set_value(mode,path,values,type=nil)
       hive, reg_path, hive_name, root_key, hive_loaded = get_reg_path_info(path)
       key_name = reg_path.join("\\")
 
-      Chef::Log.debug("Creating #{path})")
+      Chef::Log.debug("Creating #{path}")
 
       if !key_exists?(path,true)
         create_key(path)
       end
 
-      hive.send(mode, key_name, Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do |reg|
+      hive.send(mode, key_name, ::Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do |reg|
+        changed_something = false
         values.each do |k,val|
           key = "#{k}" #wtf. avoid "can't modify frozen string" in win32/registry.rb
           cur_val = nil
@@ -104,13 +105,29 @@ module Windows
           end
           if cur_val != val
             Chef::Log.debug("setting #{key}=#{val}")
-            reg[key] = val
+            
+            if type.nil?
+              type = :string
+            end
+
+            reg_type = {
+              :binary => ::Win32::Registry::REG_BINARY,
+              :string => ::Win32::Registry::REG_SZ,
+              :multi_string => ::Win32::Registry::REG_MULTI_SZ,
+              :expand_string => ::Win32::Registry::REG_EXPAND_SZ,
+              :dword => ::Win32::Registry::REG_DWORD,
+              :dword_big_endian => ::Win32::Registry::REG_DWORD_BIG_ENDIAN,
+              :qword => ::Win32::Registry::REG_QWORD
+            }[type]
+
+            reg.write(key, reg_type, val)
 
             ensure_hive_unloaded(hive_loaded)
 
-            return true
+            changed_something = true
           end
         end
+        return changed_something
       end
       return false
     end
@@ -119,7 +136,7 @@ module Windows
       hive, reg_path, hive_name, root_key, hive_loaded = get_reg_path_info(path)
       key = reg_path.join("\\")
 
-      hive.open(key, Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
+      hive.open(key, ::Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
         begin
           return reg[value]
         rescue
@@ -133,7 +150,7 @@ module Windows
     def get_values(path)
       hive, reg_path, hive_name, root_key, hive_loaded = get_reg_path_info(path)
       key = reg_path.join("\\")
-      hive.open(key, Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
+      hive.open(key, ::Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
         values = []
         begin
         reg.each_value do |name, type, data|
@@ -151,11 +168,16 @@ module Windows
       hive, reg_path, hive_name, root_key, hive_loaded = get_reg_path_info(path)
       key = reg_path.join("\\")
       Chef::Log.debug("Deleting values in #{path}")
-      hive.open(key, Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
+      hive.open(key, ::Win32::Registry::KEY_ALL_ACCESS | @@native_registry_constant) do | reg |
         values.each_key { |key|
           name = "#{key}"
-          Chef::Log.debug("Deleting value #{name} in #{path}")
-          reg.delete_value(name)
+          # Ensure delete operation is idempotent.
+          if value_exists?(path, key)
+            Chef::Log.debug("Deleting value #{name} in #{path}")
+            reg.delete_value(name)
+          else
+            Chef::Log.debug("Value #{name} in #{path} does not exist, skipping.")
+          end
         }
       end
 
@@ -178,13 +200,13 @@ module Windows
         Chef::Log.debug("Native Constant #{@@native_registry_constant}")
         Chef::Log.debug("Hive #{hive}")
 
-        hive.open(key, Win32::Registry::KEY_READ | @@native_registry_constant) do | reg |
-          begin 
+        hive.open(key, ::Win32::Registry::KEY_READ | @@native_registry_constant) do | reg |
+          begin
             rtn_value = reg[value]
             return true
           rescue
             return false
-          ensure 
+          ensure
             ensure_hive_unloaded(hive_loaded)
           end
         end
@@ -208,11 +230,11 @@ module Windows
       end
 
       begin
-        hive.open(key, Win32::Registry::Constants::KEY_READ | @@native_registry_constant )
+        hive.open(key, ::Win32::Registry::Constants::KEY_READ | @@native_registry_constant )
         return true
       rescue
         return false
-      ensure 
+      ensure
         ensure_hive_unloaded(hive_loaded)
       end
     end
@@ -229,7 +251,7 @@ module Windows
     end
 
     def resolve_user_to_sid(username)
-      begin 
+      begin
         sid = WMI::Win32_UserAccount.find(:first, :conditions => {:name => username}).sid
         Chef::Log.debug("Resolved user SID to #{sid}")
         return sid
@@ -252,7 +274,7 @@ module Windows
     end
 
     def is_user_hive?(hive)
-      if hive == Win32::Registry::HKEY_USERS
+      if hive == ::Win32::Registry::HKEY_USERS
         return true
       else
         return true
@@ -278,7 +300,7 @@ module Windows
     def load_user_hive(hive,reg_path,user_hive)
       Chef::Log.debug("Reg Path #{reg_path}")
       # See if the hive is loaded. Logged in users will have a key that is named their SID
-      # if the user has specified the a path by SID and the user is logged in, this function 
+      # if the user has specified the a path by SID and the user is logged in, this function
       # should not be executed.
       if is_user_hive?(hive) && !key_exists?("HKU\\#{user_hive}")
         Chef::Log.debug("The user is not logged in and has not been specified by SID")
